CAMEL-23648: camel-jbang - TUI refactor CamelMonitor and add threads to Memory tab
Extract StatusParser (JSON parsing) and OverviewTab (overview rendering,
sorting, sparkline charts) from CamelMonitor, reducing it from 4284 to
2674 lines. Add thread count (current/peak) to the Memory tab.
