+++ dsl/camel-jbang/camel-jbang-plugin-tui/src/main/java/org/apache/camel/dsl/jbang/core/commands/tui/CamelMonitor.java
@@ -19,14 +19,11 @@ package org.apache.camel.dsl.jbang.core.commands.tui;
 import java.io.File;
 import java.io.IOException;
 import java.io.RandomAccessFile;
-import java.net.URI;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.time.Duration;
-import java.time.Instant;
 import java.time.LocalDateTime;
-import java.time.ZoneId;
 import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -53,7 +50,6 @@ import dev.tamboui.export.ExportRequest;
 import dev.tamboui.layout.Constraint;
 import dev.tamboui.layout.Layout;
 import dev.tamboui.layout.Rect;
-import dev.tamboui.style.AnsiColor;
 import dev.tamboui.style.Color;
 import dev.tamboui.style.Style;
 import dev.tamboui.terminal.Frame;
@@ -69,9 +65,6 @@ import dev.tamboui.tui.event.KeyModifiers;
 import dev.tamboui.tui.event.PasteEvent;
 import dev.tamboui.tui.event.TickEvent;
 import dev.tamboui.widgets.Clear;
-import dev.tamboui.widgets.barchart.Bar;
-import dev.tamboui.widgets.barchart.BarChart;
-import dev.tamboui.widgets.barchart.BarGroup;
 import dev.tamboui.widgets.block.Block;
 import dev.tamboui.widgets.block.BorderType;
 import dev.tamboui.widgets.block.Title;
@@ -80,19 +73,13 @@ import dev.tamboui.widgets.list.ListState;
 import dev.tamboui.widgets.list.ListWidget;
 import dev.tamboui.widgets.list.ScrollMode;
 import dev.tamboui.widgets.paragraph.Paragraph;
-import dev.tamboui.widgets.table.Cell;
-import dev.tamboui.widgets.table.Row;
-import dev.tamboui.widgets.table.Table;
-import dev.tamboui.widgets.table.TableState;
 import dev.tamboui.widgets.tabs.Tabs;
 import dev.tamboui.widgets.tabs.TabsState;
 import org.apache.camel.dsl.jbang.core.commands.CamelCommand;
 import org.apache.camel.dsl.jbang.core.commands.CamelJBangMain;
 import org.apache.camel.dsl.jbang.core.common.CommandLineHelper;
 import org.apache.camel.dsl.jbang.core.common.PathUtils;
-import org.apache.camel.dsl.jbang.core.common.ProcessHelper;
 import org.apache.camel.dsl.jbang.core.common.VersionHelper;
-import org.apache.camel.util.TimeUtils;
 import org.apache.camel.util.json.JsonArray;
 import org.apache.camel.util.json.JsonObject;
 import org.apache.camel.util.json.Jsoner;
@@ -101,7 +88,6 @@ import picocli.CommandLine.Command;
 import sun.misc.Signal;
 
 import static org.apache.camel.dsl.jbang.core.commands.tui.MonitorContext.*;
-import static org.apache.camel.dsl.jbang.core.common.CamelCommandHelper.extractState;
 
 @Command(name = "monitor",
          description = "Live dashboard for monitoring Camel integrations",
@@ -130,9 +116,6 @@ public class CamelMonitor extends CamelCommand {
     private static final int TAB_METRICS = 8;
     private static final int TAB_MORE = 9;
 
-    // Overview sort columns
-    private static final String[] OVERVIEW_SORT_COLUMNS = { "pid", "name", "version", "status", "total", "fail" };
-
     @CommandLine.Parameters(description = "Name or pid of running Camel integration", arity = "0..1")
     String name = "*";
 
@@ -160,8 +143,6 @@ public class CamelMonitor extends CamelCommand {
     private final AtomicReference<List<InfraInfo>> infraData = new AtomicReference<>(Collections.emptyList());
     private final Map<String, VanishingInfo> vanishing = new ConcurrentHashMap<>();
     private final Map<String, VanishingInfraInfo> vanishingInfra = new ConcurrentHashMap<>();
-    private final TableState overviewTableState = new TableState();
-    private int overviewDividerIndex = -1;
     private final TabsState tabsState = new TabsState(TAB_OVERVIEW);
 
     // Sparkline: throughput history per PID (one point per second)
@@ -216,22 +197,11 @@ public class CamelMonitor extends CamelCommand {
     private final Map<String, LoadAvg> cpuLoadAvg = new ConcurrentHashMap<>();
     private final Map<String, long[]> prevCpuSample = new ConcurrentHashMap<>();
 
-    // Overview sort state
-    private String overviewSort = "name";
-    private int overviewSortIndex = 1;
-    private boolean overviewSortReversed;
-
     // Trace/history data — shared between CamelMonitor and tabs
     private final AtomicReference<List<TraceEntry>> traces = new AtomicReference<>(Collections.emptyList());
     private final Map<String, Long> traceFilePositions = new ConcurrentHashMap<>();
 
     // selectedPid is stored on ctx (MonitorContext) so tabs can access it
-
-    // Overview chart mode
-    private static final int CHART_ALL = 0;
-    private static final int CHART_SINGLE = 1;
-    private static final int CHART_OFF = 2;
-    private int chartMode = CHART_SINGLE;
 
     private volatile long lastRefresh;
     private boolean showKillConfirm;
@@ -291,6 +261,7 @@ public class CamelMonitor extends CamelCommand {
     private InflightTab inflightTab;
     private MemoryTab memoryTab;
     private ThreadsTab threadsTab;
+    private OverviewTab overviewTab;
 
     // "Switch integration" popup state
     private boolean showSwitchPopup;
@@ -359,6 +330,9 @@ public class CamelMonitor extends CamelCommand {
         inflightTab = new InflightTab(ctx);
         memoryTab = new MemoryTab(ctx, heapMemHistory);
         threadsTab = new ThreadsTab(ctx);
+        overviewTab = new OverviewTab(
+                ctx, throughputHistory, failedHistory, cpuLoadAvg,
+                this::resetIntegrationTabState);
 
         // Initial data load (synchronous before TUI starts)
         refreshDataSync();
@@ -478,7 +452,7 @@ public class CamelMonitor extends CamelCommand {
                             default -> null;
                         };
                         if (activeMoreTab != null) {
-                            selectCurrentIntegration();
+                            overviewTab.selectCurrentIntegration();
                             tabsState.select(TAB_MORE);
                             activeMoreTab.onTabSelected();
                         }
@@ -600,7 +574,7 @@ public class CamelMonitor extends CamelCommand {
                 } else {
                     int prev = (tabsState.selected() - 1 + NUM_TABS) % NUM_TABS;
                     if (prev != TAB_OVERVIEW) {
-                        selectCurrentIntegration();
+                        overviewTab.selectCurrentIntegration();
                     }
                     tabsState.select(prev);
                 }
@@ -613,7 +587,7 @@ public class CamelMonitor extends CamelCommand {
                 } else {
                     int next = (tabsState.selected() + 1) % NUM_TABS;
                     if (next != TAB_OVERVIEW) {
-                        selectCurrentIntegration();
+                        overviewTab.selectCurrentIntegration();
                     }
                     tabsState.select(next);
                 }
@@ -705,29 +679,13 @@ public class CamelMonitor extends CamelCommand {
 
             // Enter to drill into selected integration
             if (ke.isConfirm() && tab == TAB_OVERVIEW) {
-                selectCurrentIntegration();
+                overviewTab.selectCurrentIntegration();
                 if (ctx.selectedPid != null) {
                     tabsState.select(TAB_LOG);
                 }
                 return true;
             }
 
-            // Overview tab: sort
-            if (tab == TAB_OVERVIEW && ke.isChar('s')) {
-                overviewSortIndex = (overviewSortIndex + 1) % OVERVIEW_SORT_COLUMNS.length;
-                overviewSort = OVERVIEW_SORT_COLUMNS[overviewSortIndex];
-                overviewSortReversed = false;
-                return true;
-            }
-            if (tab == TAB_OVERVIEW && ke.isChar('S')) {
-                overviewSortReversed = !overviewSortReversed;
-                return true;
-            }
-            // Overview tab: cycle chart between all integrations, selected only, and off
-            if (tab == TAB_OVERVIEW && ke.isCharIgnoreCase('a')) {
-                chartMode = (chartMode + 1) % 3;
-                return true;
-            }
             // Overview tab: start/stop all routes for selected integration (not infra)
             if (tab == TAB_OVERVIEW && ke.isChar('p') && ctx.selectedPid != null && !isInfraSelected()) {
                 IntegrationInfo selInfo = findSelectedIntegration();
@@ -859,7 +817,7 @@ public class CamelMonitor extends CamelCommand {
 
     private boolean handleTabKey(int tab) {
         if (tab != TAB_OVERVIEW) {
-            selectCurrentIntegration();
+            overviewTab.selectCurrentIntegration();
         }
         if (tab == TAB_LOG) {
             logTab.onTabSelected();
@@ -895,34 +853,6 @@ public class CamelMonitor extends CamelCommand {
         return true;
     }
 
-    // Returns integrations in the same order the overview table renders them.
-    // Must be used anywhere that translates a table row index to a PID.
-    private List<IntegrationInfo> sortedOverviewInfos() {
-        List<IntegrationInfo> infos = new ArrayList<>(data.get());
-        infos.sort(this::sortOverview);
-        return infos;
-    }
-
-    private void selectCurrentIntegration() {
-        if (ctx.selectedPid != null) {
-            if (findSelectedIntegration() != null || findSelectedInfra() != null) {
-                return;
-            }
-            ctx.selectedPid = null;
-        }
-        List<IntegrationInfo> infos = sortedOverviewInfos();
-        List<InfraInfo> infras = infraData.get();
-        Integer sel = overviewTableState.selected();
-        if (sel != null && sel >= 0) {
-            String pid = overviewIndexToPid(infos, infras, sel);
-            if (pid != null) {
-                ctx.selectedPid = pid;
-            }
-        } else if (infos.size() == 1) {
-            ctx.selectedPid = infos.get(0).pid;
-        }
-    }
-
     private List<Long> selectedPidAsList() {
         if (ctx.selectedPid == null) {
             return Collections.emptyList();
@@ -932,35 +862,6 @@ public class CamelMonitor extends CamelCommand {
         } catch (NumberFormatException e) {
             return Collections.emptyList();
         }
-    }
-
-    private void syncSelectedPid() {
-        List<IntegrationInfo> infos = sortedOverviewInfos();
-        List<InfraInfo> infras = infraData.get();
-        Integer sel = overviewTableState.selected();
-        String newPid = null;
-        if (sel != null && sel >= 0) {
-            newPid = overviewIndexToPid(infos, infras, sel);
-        }
-        if (newPid == null && infos.size() == 1) {
-            newPid = infos.get(0).pid;
-        }
-        if (newPid != null && !newPid.equals(ctx.selectedPid)) {
-            ctx.selectedPid = newPid;
-            ctx.lastSelectedName = null;
-            resetIntegrationTabState();
-        }
-    }
-
-    private String overviewIndexToPid(List<IntegrationInfo> infos, List<InfraInfo> infras, int index) {
-        if (index < infos.size()) {
-            return infos.get(index).pid;
-        }
-        int infraIndex = index - infos.size() - (overviewDividerIndex >= 0 ? 1 : 0);
-        if (infraIndex >= 0 && infraIndex < infras.size()) {
-            return infras.get(infraIndex).pid;
-        }
-        return null;
     }
 
     private void resetIntegrationTabState() {
@@ -979,40 +880,11 @@ public class CamelMonitor extends CamelCommand {
     }
 
     private void navigateUp() {
-        MonitorTab tab = activeTab();
-        if (tab != null) {
-            tab.navigateUp();
-        } else {
-            overviewTableState.selectPrevious();
-            // Skip the divider row
-            Integer sel = overviewTableState.selected();
-            if (sel != null && overviewDividerIndex >= 0 && sel == overviewDividerIndex) {
-                overviewTableState.selectPrevious();
-            }
-            syncSelectedPid();
-        }
+        activeTab().navigateUp();
     }
 
     private void navigateDown() {
-        MonitorTab tab = activeTab();
-        if (tab != null) {
-            tab.navigateDown();
-        } else {
-            int totalRows = overviewTotalRows();
-            overviewTableState.selectNext(totalRows);
-            // Skip the divider row
-            Integer sel = overviewTableState.selected();
-            if (sel != null && overviewDividerIndex >= 0 && sel == overviewDividerIndex) {
-                overviewTableState.selectNext(totalRows);
-            }
-            syncSelectedPid();
-        }
-    }
-
-    private int overviewTotalRows() {
-        int integrations = sortedOverviewInfos().size();
-        int infra = infraData.get().size();
-        return integrations + (infra > 0 ? 1 : 0) + infra;
+        activeTab().navigateDown();
     }
 
     // ---- Rendering ----
@@ -1250,11 +1122,7 @@ public class CamelMonitor extends CamelCommand {
         // switching tabs if TamboUI's buffer diff does not reset every cell in the region.
         frame.buffer().clear(area);
         MonitorTab tab = activeTab();
-        if (tab != null) {
-            tab.render(frame, area);
-        } else {
-            renderOverview(frame, area);
-        }
+        tab.render(frame, area);
         // Render "More" popup overlay when visible
         if (showMorePopup) {
             renderMorePopup(frame, area);
@@ -1400,6 +1268,7 @@ public class CamelMonitor extends CamelCommand {
 
     private MonitorTab activeTab() {
         return switch (tabsState.selected()) {
+            case TAB_OVERVIEW -> overviewTab;
             case TAB_LOG -> logTab;
             case TAB_ROUTES -> routesTab;
             case TAB_ENDPOINTS -> endpointsTab;
@@ -1411,517 +1280,6 @@ public class CamelMonitor extends CamelCommand {
             case TAB_MORE -> activeMoreTab;
             default -> null;
         };
-    }
-
-    // ---- Tab 1: Overview ----
-
-    private void renderOverview(Frame frame, Rect area) {
-        List<IntegrationInfo> infos = sortedOverviewInfos();
-        List<InfraInfo> infraInfos = infraData.get();
-
-        // Build unified row list: integrations, then divider, then infra
-        int integrationCount = infos.size();
-        int infraCount = infraInfos.size();
-        overviewDividerIndex = infraCount > 0 ? integrationCount : -1;
-
-        // Keep the table selection index tracking the same PID across sort changes and data refreshes
-        if (ctx.selectedPid != null) {
-            for (int i = 0; i < infos.size(); i++) {
-                if (ctx.selectedPid.equals(infos.get(i).pid)) {
-                    overviewTableState.select(i);
-                    break;
-                }
-            }
-            for (int i = 0; i < infraInfos.size(); i++) {
-                if (ctx.selectedPid.equals(infraInfos.get(i).pid)) {
-                    int tableIndex = integrationCount + (overviewDividerIndex >= 0 ? 1 : 0) + i;
-                    overviewTableState.select(tableIndex);
-                    break;
-                }
-            }
-        }
-
-        // Split: table + chart or info panel
-        boolean hasSparkline = chartMode != CHART_OFF && !throughputHistory.isEmpty() && !isInfraSelected();
-        boolean showInfoPanel = isInfraSelected() && findSelectedInfra() != null && !hasSparkline;
-        List<Constraint> constraints = new ArrayList<>();
-        constraints.add(Constraint.fill());
-        if (hasSparkline) {
-            constraints.add(Constraint.length(14));
-        } else if (showInfoPanel) {
-            constraints.add(Constraint.length(10));
-        }
-        List<Rect> chunks = Layout.vertical()
-                .constraints(constraints)
-                .split(area);
-
-        // Integration table
-        List<Row> rows = new ArrayList<>();
-        for (IntegrationInfo info : infos) {
-            if (info.vanishing) {
-                long elapsed = System.currentTimeMillis() - info.vanishStart;
-                float fade = 1.0f - Math.min(1.0f, (float) elapsed / VANISH_DURATION_MS);
-                int gray = (int) (100 * fade);
-                Style dimStyle = Style.EMPTY.fg(Color.indexed(232 + Math.min(gray / 4, 23)));
-
-                String vanishName = "\ud83d\udc2a " + (info.name != null ? info.name : "");
-                rows.add(Row.from(
-                        Cell.from(Span.styled(info.pid, dimStyle)),
-                        Cell.from(Span.styled(vanishName, dimStyle)),
-                        Cell.from(Span.styled("", dimStyle)),
-                        Cell.from(Span.styled("", dimStyle)),
-                        Cell.from(Span.styled("\u2716 Stopped", Style.EMPTY.fg(Color.LIGHT_RED).dim())),
-                        Cell.from(Span.styled(info.ago != null ? info.ago : "", dimStyle)),
-                        Cell.from(Span.styled("", dimStyle)),
-                        Cell.from(Span.styled("", dimStyle)),
-                        Cell.from(Span.styled("", dimStyle)),
-                        Cell.from(Span.styled("", dimStyle)),
-                        Cell.from(Span.styled("", dimStyle)),
-                        Cell.from(Span.styled("", dimStyle))));
-            } else {
-                Style statusStyle = switch (extractState(info.state)) {
-                    case "Started", "Running" -> Style.EMPTY.fg(Color.GREEN);
-                    case "Stopped" -> Style.EMPTY.fg(Color.LIGHT_RED);
-                    default -> Style.EMPTY.fg(Color.YELLOW);
-                };
-
-                Style failStyle = info.failed > 0 ? Style.EMPTY.fg(Color.LIGHT_RED).bold() : Style.EMPTY;
-
-                String sinceLastDisplay = formatSinceLast(info);
-
-                String nameText = "🐪 " + (info.name != null ? info.name : "");
-                Line nameLine = info.devMode
-                        ? Line.from(
-                                Span.styled(nameText, Style.EMPTY.fg(Color.CYAN)),
-                                Span.styled(" [dev]", Style.EMPTY.fg(Color.YELLOW).dim()))
-                        : Line.from(Span.styled(nameText, Style.EMPTY.fg(Color.CYAN)));
-                rows.add(Row.from(
-                        Cell.from(info.pid),
-                        Cell.from(nameLine),
-                        Cell.from(info.camelVersion != null ? info.camelVersion : ""),
-                        centerCell(info.ready != null ? info.ready : "", 5),
-                        Cell.from(Span.styled(extractState(info.state), statusStyle)),
-                        Cell.from(info.ago != null ? info.ago : ""),
-                        rightCell(info.routeStarted + "/" + info.routeTotal, 7),
-                        rightCell(info.throughput != null ? info.throughput : "", 8),
-                        rightCell(String.valueOf(info.exchangesTotal), 8),
-                        rightCell(String.valueOf(info.failed), 6, failStyle),
-                        rightCell(String.valueOf(info.inflight), 8),
-                        Cell.from(sinceLastDisplay)));
-            }
-        }
-
-        Row header = Row.from(
-                Cell.from(Span.styled(overviewSortLabel("PID", "pid"), overviewSortStyle("pid"))),
-                Cell.from(Span.styled(overviewSortLabel("NAME", "name"), overviewSortStyle("name"))),
-                Cell.from(Span.styled(overviewSortLabel("VERSION", "version"), overviewSortStyle("version"))),
-                centerCell("READY", 5, Style.EMPTY.bold()),
-                Cell.from(Span.styled(overviewSortLabel("STATUS", "status"), overviewSortStyle("status"))),
-                Cell.from(Span.styled("AGE", Style.EMPTY.bold())),
-                rightCell("ROUTE", 7, Style.EMPTY.bold()),
-                rightCell("MSG/S", 8, Style.EMPTY.bold()),
-                rightCell(overviewSortLabel("TOTAL", "total"), 8, overviewSortStyle("total")),
-                rightCell(overviewSortLabel("FAIL", "fail"), 6, overviewSortStyle("fail")),
-                rightCell("INFLIGHT", 8, Style.EMPTY.bold()),
-                Cell.from(Span.styled("SINCE-LAST", Style.EMPTY.bold())));
-
-        // Divider row between integrations and infra services
-        if (overviewDividerIndex >= 0) {
-            rows.add(Row.from(
-                    Cell.from(""),
-                    Cell.from(Span.styled("─── Infra Services ───", Style.EMPTY.dim())),
-                    Cell.from(""), Cell.from(""), Cell.from(""), Cell.from(""),
-                    Cell.from(""), Cell.from(""), Cell.from(""), Cell.from(""),
-                    Cell.from(""), Cell.from("")));
-        }
-
-        // Infra rows adapted to 12-column layout
-        for (InfraInfo info : infraInfos) {
-            if (info.vanishing) {
-                long elapsed = System.currentTimeMillis() - info.vanishStart;
-                float fade = 1.0f - Math.min(1.0f, (float) elapsed / VANISH_DURATION_MS);
-                int gray = (int) (100 * fade);
-                Style dimStyle = Style.EMPTY.fg(Color.indexed(232 + Math.min(gray / 4, 23)));
-                String vanishAlias = "🔧  " + info.alias;
-                rows.add(Row.from(
-                        Cell.from(Span.styled(info.pid, dimStyle)),
-                        Cell.from(Span.styled(vanishAlias, dimStyle)),
-                        Cell.from(Span.styled("", dimStyle)),
-                        Cell.from(Span.styled("", dimStyle)),
-                        Cell.from(Span.styled("✖ Stopped", Style.EMPTY.fg(Color.LIGHT_RED).dim())),
-                        Cell.from(Span.styled("", dimStyle)),
-                        Cell.from(Span.styled("", dimStyle)),
-                        Cell.from(Span.styled("", dimStyle)),
-                        Cell.from(Span.styled("", dimStyle)),
-                        Cell.from(Span.styled("", dimStyle)),
-                        Cell.from(Span.styled("", dimStyle)),
-                        Cell.from(Span.styled("", dimStyle))));
-            } else {
-                Style statusStyle = info.alive ? Style.EMPTY.fg(Color.GREEN) : Style.EMPTY.fg(Color.LIGHT_RED);
-                String statusText = info.alive ? "Running" : "Stopped";
-                String infraAlias = "🔧  " + info.alias;
-                String version = info.serviceVersion != null ? info.serviceVersion : "";
-                rows.add(Row.from(
-                        Cell.from(info.pid),
-                        Cell.from(Span.styled(infraAlias, Style.EMPTY.fg(Color.MAGENTA))),
-                        Cell.from(version),
-                        Cell.from(""),
-                        Cell.from(Span.styled(statusText, statusStyle)),
-                        Cell.from(""),
-                        Cell.from(""),
-                        Cell.from(""),
-                        Cell.from(""),
-                        Cell.from(""),
-                        Cell.from(""),
-                        Cell.from("")));
-            }
-        }
-
-        Style overviewHighlight = Style.EMPTY.fg(Color.WHITE).bold().onBlue();
-        Table table = Table.builder()
-                .rows(rows)
-                .header(header)
-                .widths(
-                        Constraint.length(8),
-                        Constraint.fill(),
-                        Constraint.length(16),
-                        Constraint.length(5),
-                        Constraint.length(10),
-                        Constraint.length(8),
-                        Constraint.length(7),
-                        Constraint.length(8),
-                        Constraint.length(8),
-                        Constraint.length(6),
-                        Constraint.length(8),
-                        Constraint.length(12))
-                .highlightStyle(overviewHighlight)
-                .highlightSpacing(Table.HighlightSpacing.ALWAYS)
-                .block(Block.builder().borderType(BorderType.ROUNDED).title(" Overview ").build())
-                .build();
-
-        frame.renderStatefulWidget(table, chunks.get(0), overviewTableState);
-
-        // Split green/red throughput bar chart with Y and X axes
-        if (hasSparkline && chunks.size() > 1) {
-            Rect chartTotalArea = chunks.get(chunks.size() - 1);
-
-            // Split chart area horizontally: bar chart (fill) + info panel (30 cols)
-            List<Rect> chartHSplit = Layout.horizontal()
-                    .constraints(Constraint.fill(), Constraint.length(30))
-                    .split(chartTotalArea);
-            Rect chartArea = chartHSplit.get(0);
-            Rect infoArea = chartHSplit.get(1);
-
-            // Split chart area: chart rows (13) + x-axis label row (1)
-            List<Rect> vChunks = Layout.vertical()
-                    .constraints(Constraint.fill(), Constraint.length(1))
-                    .split(chartArea);
-
-            // Split chart rows: y-axis labels (4 cols) + bar chart (fill)
-            List<Rect> hChunks = Layout.horizontal()
-                    .constraints(Constraint.length(4), Constraint.fill())
-                    .split(vChunks.get(0));
-
-            Rect barChartArea = hChunks.get(1);
-
-            // Compute how many ticks fit: each tick = 2 bars × barWidth=1 = 2 cols
-            int innerBarCols = Math.max(2, barChartArea.width() - 2); // minus block borders
-            int renderPoints = Math.min(MAX_SPARKLINE_POINTS, innerBarCols / 2);
-
-            // Merge throughput histories: all PIDs or selected only
-            long[] mergedTotal = new long[renderPoints];
-            long[] mergedFailed = new long[renderPoints];
-            String chartPid = (chartMode == CHART_SINGLE && ctx.selectedPid != null) ? ctx.selectedPid : null;
-            for (int i = 0; i < renderPoints; i++) {
-                for (Map.Entry<String, LinkedList<Long>> e : throughputHistory.entrySet()) {
-                    if (chartPid == null || chartPid.equals(e.getKey())) {
-                        int idx = e.getValue().size() - renderPoints + i;
-                        if (idx >= 0) {
-                            mergedTotal[i] += e.getValue().get(idx);
-                        }
-                    }
-                }
-                for (Map.Entry<String, LinkedList<Long>> e : failedHistory.entrySet()) {
-                    if (chartPid == null || chartPid.equals(e.getKey())) {
-                        int idx = e.getValue().size() - renderPoints + i;
-                        if (idx >= 0) {
-                            mergedFailed[i] += e.getValue().get(idx);
-                        }
-                    }
-                }
-            }
-
-            long maxTp = 0;
-            for (long v : mergedTotal) {
-                maxTp = Math.max(maxTp, v);
-            }
-            long curTp = mergedTotal[renderPoints - 1];
-            long curFailed = mergedFailed[renderPoints - 1];
-            long curOk = Math.max(0, curTp - curFailed);
-
-            // Styled legend in chart title
-            Line titleLine;
-            if (chartMode == CHART_SINGLE && ctx.selectedPid != null) {
-                IntegrationInfo chartSel = findSelectedIntegration();
-                String chartName = chartSel != null ? TuiHelper.truncate(chartSel.name, 12) : ctx.selectedPid;
-                titleLine = Line.from(
-                        Span.raw(" ["),
-                        Span.styled(chartName, Style.EMPTY.fg(Color.YELLOW)),
-                        Span.raw(String.format("] Throughput: %d msg/s  ", curTp)),
-                        Span.styled("■", Style.EMPTY.fg(Color.ansi(AnsiColor.BRIGHT_GREEN))),
-                        Span.raw(String.format(" ok:%d  ", curOk)),
-                        Span.styled("■", Style.EMPTY.fg(Color.RED)),
-                        Span.raw(String.format(" fail:%d ", curFailed)));
-            } else {
-                titleLine = Line.from(
-                        Span.raw(String.format(" [All] Throughput: %d msg/s  ", curTp)),
-                        Span.styled("■", Style.EMPTY.fg(Color.ansi(AnsiColor.BRIGHT_GREEN))),
-                        Span.raw(String.format(" ok:%d  ", curOk)),
-                        Span.styled("■", Style.EMPTY.fg(Color.RED)),
-                        Span.raw(String.format(" fail:%d ", curFailed)));
-            }
-
-            // Build bar groups (ok=bright green, failed=red), no bar value labels
-            List<BarGroup> groups = new ArrayList<>();
-            for (int i = 0; i < renderPoints; i++) {
-                long failed = Math.min(mergedFailed[i], mergedTotal[i]);
-                long ok = Math.max(0, mergedTotal[i] - failed);
-                groups.add(BarGroup.of(
-                        Bar.builder().value(ok).textValue("").style(Style.EMPTY.fg(Color.ansi(AnsiColor.BRIGHT_GREEN))).build(),
-                        Bar.builder().value(failed).textValue("").style(Style.EMPTY.fg(Color.RED)).build()));
-            }
-
-            BarChart barChart = BarChart.builder()
-                    .data(groups)
-                    .max(maxTp > 0 ? maxTp + 2 : 2)
-                    .barWidth(1)
-                    .barGap(0)
-                    .groupGap(0)
-                    .block(Block.builder().borderType(BorderType.ROUNDED)
-                            .title(Title.from(titleLine)).build())
-                    .build();
-
-            frame.renderWidget(barChart, barChartArea);
-
-            // Y-axis: scale labels aligned with bar chart inner rows
-            int barRows = vChunks.get(0).height() - 2; // minus top + bottom border
-            List<Line> yLines = new ArrayList<>();
-            Style dimStyle = Style.EMPTY.dim();
-            for (int row = 0; row < vChunks.get(0).height(); row++) {
-                int barRow = row - 1; // bar area starts after top border
-                if (barRow == 0) {
-                    yLines.add(Line.from(Span.styled(String.format("%3d", maxTp), dimStyle)));
-                } else if (barRows > 4 && barRow == barRows / 2) {
-                    yLines.add(Line.from(Span.styled(String.format("%3d", maxTp / 2), dimStyle)));
-                } else if (barRow == barRows - 1) {
-                    yLines.add(Line.from(Span.styled("  0", dimStyle)));
-                } else {
-                    yLines.add(Line.from(""));
-                }
-            }
-            frame.renderWidget(Paragraph.builder().text(Text.from(yLines)).build(), hChunks.get(0));
-
-            // X-axis: time labels drawn into the bottom row
-            if (!vChunks.get(1).isEmpty()) {
-                int barInnerStartX = barChartArea.x() + 1; // inside left border
-                int xAxisY = vChunks.get(1).y();
-                // Markers at: oldest, 1/4, 1/2, 3/4, newest
-                int[][] markerIndices = {
-                        { 0, renderPoints },
-                        { renderPoints / 4, renderPoints - renderPoints / 4 },
-                        { renderPoints / 2, renderPoints / 2 },
-                        { 3 * renderPoints / 4, renderPoints / 4 },
-                        { renderPoints - 1, 0 }
-                };
-                for (int[] m : markerIndices) {
-                    int groupIdx = m[0];
-                    int secsAgo = m[1];
-                    String label = secsAgo == 0 ? "now" : "-" + secsAgo + "s";
-                    int markerX = barInnerStartX + groupIdx * 2;
-                    if (markerX + label.length() <= barChartArea.right()) {
-                        frame.buffer().setString(markerX, xAxisY, label, dimStyle);
-                    }
-                }
-            }
-
-            // Info panel: heap and threads for the selected integration
-            renderOverviewInfoPanel(frame, infoArea);
-        } else if (showInfoPanel) {
-            renderOverviewInfoPanel(frame, chunks.get(chunks.size() - 1));
-        }
-    }
-
-    private void renderOverviewInfoPanel(Frame frame, Rect area) {
-        // Check if an infra service is selected — show connection details instead
-        InfraInfo infraSel = findSelectedInfra();
-        if (infraSel != null) {
-            renderInfraInfoPanel(frame, area, infraSel);
-            return;
-        }
-
-        IntegrationInfo sel = findSelectedIntegration();
-        // Fall back to the single active integration when nothing is explicitly selected
-        if (sel == null) {
-            List<IntegrationInfo> active = data.get().stream().filter(i -> !i.vanishing).toList();
-            if (active.size() == 1) {
-                sel = active.get(0);
-            }
-        }
-        Block infoBlock = Block.builder().borderType(BorderType.ROUNDED).build();
-        frame.renderWidget(infoBlock, area);
-        Rect inner = infoBlock.inner(area);
-        List<Line> lines = new ArrayList<>();
-        Style dim = Style.EMPTY.dim();
-        if (sel != null) {
-            // Identity
-            if (sel.platform != null) {
-                String plat = sel.platformVersion != null
-                        ? sel.platform + " v" + sel.platformVersion
-                        : sel.platform;
-                lines.add(Line.from(
-                        Span.styled("Runtime: ", dim),
-                        Span.raw(TuiHelper.truncate(plat, inner.width() - 9))));
-            }
-            if (sel.camelVersion != null) {
-                lines.add(Line.from(
-                        Span.styled("Version: ", dim),
-                        Span.raw(TuiHelper.truncate(sel.camelVersion, inner.width() - 9))));
-            }
-            if (sel.profile != null || sel.reloaded > 0) {
-                List<Span> profileSpans = new ArrayList<>();
-                if (sel.profile != null) {
-                    profileSpans.add(Span.styled("Profile: ", dim));
-                    profileSpans.add(Span.raw(sel.profile));
-                }
-                if (sel.reloaded > 0) {
-                    if (!profileSpans.isEmpty()) {
-                        profileSpans.add(Span.raw("    "));
-                    }
-                    profileSpans.add(Span.styled("Reload: ", dim));
-                    profileSpans.add(Span.raw(String.valueOf(sel.reloaded)));
-                }
-                lines.add(Line.from(profileSpans));
-            }
-            lines.add(Line.from(Span.raw("")));
-            // Resources
-            if (sel.javaVersion != null) {
-                lines.add(Line.from(
-                        Span.styled("JVM:  ", dim),
-                        Span.raw(TuiHelper.truncate(sel.javaVersion, inner.width() - 6))));
-            }
-            if (sel.javaVendor != null) {
-                lines.add(Line.from(
-                        Span.styled("      ", dim),
-                        Span.raw(TuiHelper.truncate(sel.javaVendor, inner.width() - 6))));
-            }
-            if (sel.javaVmName != null) {
-                lines.add(Line.from(
-                        Span.styled("      ", dim),
-                        Span.raw(TuiHelper.truncate(sel.javaVmName, inner.width() - 6))));
-            }
-            lines.add(Line.from(
-                    Span.styled("Uptime: ", dim),
-                    Span.raw(sel.ago != null ? sel.ago : "-")));
-            if (sel.heapMemUsed > 0) {
-                String heap = formatMemory(sel.heapMemUsed, sel.heapMemMax);
-                long pct = sel.heapMemMax > 0 ? sel.heapMemUsed * 100 / sel.heapMemMax : 0;
-                lines.add(Line.from(
-                        Span.styled("Heap: ", dim),
-                        Span.raw(heap + " " + pct + "%")));
-            }
-            if (sel.nonHeapMemUsed > 0) {
-                lines.add(Line.from(
-                        Span.styled("Meta: ", dim),
-                        Span.raw(formatMemory(sel.nonHeapMemUsed, 0))));
-            }
-            if (sel.threadCount > 0) {
-                lines.add(Line.from(
-                        Span.styled("Thds: ", dim),
-                        Span.raw(sel.threadCount + " / " + sel.peakThreadCount)));
-            }
-            LoadAvg cpu = cpuLoadAvg.get(sel.pid);
-            boolean hasInfl = sel.inflightLoad01 != null && !sel.inflightLoad01.isEmpty();
-            if (cpu != null || hasInfl) {
-                lines.add(Line.from(Span.raw("")));
-                lines.add(Line.from(Span.styled("Load (1m/5m/15m):", dim)));
-                if (cpu != null) {
-                    lines.add(Line.from(
-                            Span.styled("CPU:  ", dim),
-                            Span.raw(cpu.format("%.1f / %.1f / %.1f %%"))));
-                }
-                if (hasInfl) {
-                    lines.add(Line.from(
-                            Span.styled("Infl: ", dim),
-                            Span.raw(sel.inflightLoad01 + " / " + sel.inflightLoad05 + " / " + sel.inflightLoad15)));
-                }
-            }
-        } else {
-            lines.add(Line.from(Span.raw("-")));
-        }
-        frame.renderWidget(Paragraph.builder().text(Text.from(lines)).build(), inner);
-    }
-
-    private void renderInfraInfoPanel(Frame frame, Rect area, InfraInfo infra) {
-        Block infoBlock = Block.builder().borderType(BorderType.ROUNDED).build();
-        frame.renderWidget(infoBlock, area);
-        Rect inner = infoBlock.inner(area);
-        List<Line> lines = new ArrayList<>();
-        Style dim = Style.EMPTY.dim();
-        lines.add(Line.from(
-                Span.styled("Service: ", dim),
-                Span.styled(infra.alias, Style.EMPTY.fg(Color.MAGENTA))));
-        lines.add(Line.from(Span.raw("")));
-        // Show connection properties with cleaned-up key names
-        for (Map.Entry<String, Object> e : infra.properties.entrySet()) {
-            String key = e.getKey();
-            // Strip "get" prefix and capitalize
-            if (key.startsWith("get") && key.length() > 3) {
-                key = key.substring(3);
-            }
-            String value = String.valueOf(e.getValue());
-            lines.add(Line.from(
-                    Span.styled(key + ": ", dim),
-                    Span.raw(TuiHelper.truncate(value, inner.width() - key.length() - 2))));
-        }
-        frame.renderWidget(Paragraph.builder().text(Text.from(lines)).build(), inner);
-    }
-
-    // ---- Overview helpers ----
-
-    private int sortOverview(IntegrationInfo a, IntegrationInfo b) {
-        if (a.vanishing != b.vanishing) {
-            return a.vanishing ? 1 : -1;
-        }
-        int result = switch (overviewSort) {
-            case "pid" -> {
-                String pa = a.pid != null ? a.pid : "";
-                String pb = b.pid != null ? b.pid : "";
-                yield pa.compareTo(pb);
-            }
-            case "name" -> {
-                String na = a.name != null ? a.name : "";
-                String nb = b.name != null ? b.name : "";
-                yield na.compareToIgnoreCase(nb);
-            }
-            case "version" -> {
-                String va = a.camelVersion != null ? a.camelVersion : "";
-                String vb = b.camelVersion != null ? b.camelVersion : "";
-                yield va.compareToIgnoreCase(vb);
-            }
-            case "status" -> Integer.compare(a.state, b.state);
-            case "total" -> Long.compare(b.exchangesTotal, a.exchangesTotal);
-            case "fail" -> Long.compare(b.failed, a.failed);
-            default -> 0;
-        };
-        return overviewSortReversed ? -result : result;
-    }
-
-    private String overviewSortLabel(String label, String column) {
-        return sortLabel(label, column, overviewSort, overviewSortReversed);
-    }
-
-    private Style overviewSortStyle(String column) {
-        return sortStyle(column, overviewSort);
     }
 
     private void stopSelectedProcess(boolean forceKill) {
@@ -2194,9 +1552,10 @@ public class CamelMonitor extends CamelCommand {
         } else {
             MonitorTab tab = activeTab();
 
-            if (tab != null) {
+            if (tabsState.selected() == TAB_OVERVIEW) {
+                renderOverviewFooter(spans);
+            } else {
                 tab.renderFooter(spans);
-                // Insert F2/F3 after the first hint (Esc) — each hint is 2 spans (key + label)
                 int insertPos = Math.min(2, spans.size());
                 List<Span> fKeySpans = new ArrayList<>();
                 hint(fKeySpans, "F2", "actions");
@@ -2204,8 +1563,6 @@ public class CamelMonitor extends CamelCommand {
                     hint(fKeySpans, "F3", "switch");
                 }
                 spans.addAll(insertPos, fKeySpans);
-            } else {
-                renderOverviewFooter(spans);
             }
         }
 
@@ -2264,23 +1621,16 @@ public class CamelMonitor extends CamelCommand {
             actionsPopup.renderFooter(spans);
             return;
         }
-        hint(spans, "q", "quit");
-        hint(spans, "F2", "actions");
+        overviewTab.renderFooter(spans);
+        // Insert F2/F3 after first hint (q) — each hint is 2 spans (key + label)
+        int insertPos = Math.min(2, spans.size());
+        List<Span> fKeySpans = new ArrayList<>();
+        hint(fKeySpans, "F2", "actions");
         if (getNonVanishingIntegrations().size() > 1) {
-            hint(spans, "F3", "switch");
+            hint(fKeySpans, "F3", "switch");
         }
-        if (ctx.selectedPid != null) {
-            hint(spans, "Esc", "unselect");
-        }
-        hint(spans, "↑↓", "navigate");
-        if (!isInfraSelected()) {
-            hint(spans, "s", "sort");
-            hint(spans, "a", "chart " + switch (chartMode) {
-                case CHART_ALL -> "[all]";
-                case CHART_SINGLE -> "[single]";
-                default -> "[off]";
-            });
-        }
+        spans.addAll(insertPos, fKeySpans);
+        // Process action hints
         if (ctx.selectedPid != null && !isInfraSelected()) {
             IntegrationInfo selInfo = findSelectedIntegration();
             if (selInfo != null) {
@@ -2294,7 +1644,6 @@ public class CamelMonitor extends CamelCommand {
             hint(spans, "x", "stop");
             hint(spans, "X", "kill");
         }
-        hint(spans, isInfraSelected() ? "1-2" : "1-9", "tabs");
     }
 
     // ---- Data Loading ----
@@ -2328,7 +1677,7 @@ public class CamelMonitor extends CamelCommand {
                     .forEach(ph -> {
                         JsonObject root = loadStatus(ph.pid());
                         if (root != null) {
-                            IntegrationInfo info = parseIntegration(ph, root);
+                            IntegrationInfo info = StatusParser.parseIntegration(ph, root);
                             if (info != null) {
                                 infos.add(info);
                                 updateThroughputHistory(info);
@@ -2449,7 +1798,7 @@ public class CamelMonitor extends CamelCommand {
                 List<InfraInfo> infras = infraData.get();
                 if (!infras.isEmpty()) {
                     int firstInfraIndex = infos.size() + (infras.size() > 0 ? 1 : 0);
-                    overviewTableState.select(firstInfraIndex);
+                    overviewTab.tableState.select(firstInfraIndex);
                     ctx.selectedPid = infras.get(0).pid;
                 }
             }
@@ -2559,7 +1908,7 @@ public class CamelMonitor extends CamelCommand {
                         } catch (Exception e) {
                             // ignore parse errors
                         }
-                        info.serviceVersion = objToString(info.properties.get("serviceVersion"));
+                        info.serviceVersion = StatusParser.objToString(info.properties.get("serviceVersion"));
                         infraInfos.add(info);
                     }
                 }
@@ -2869,7 +2218,7 @@ public class CamelMonitor extends CamelCommand {
                     if (tracesArray instanceof List<?> traceList) {
                         for (Object traceObj : traceList) {
                             if (traceObj instanceof JsonObject traceJson) {
-                                TraceEntry entry = parseTraceEntry(traceJson, pid);
+                                TraceEntry entry = StatusParser.parseTraceEntry(traceJson, pid);
                                 if (entry != null) {
                                     allTraces.add(entry);
                                 }
@@ -2877,7 +2226,7 @@ public class CamelMonitor extends CamelCommand {
                         }
                     } else {
                         // Fallback: try parsing the line itself as a trace entry
-                        TraceEntry entry = parseTraceEntry(json, pid);
+                        TraceEntry entry = StatusParser.parseTraceEntry(json, pid);
                         if (entry != null) {
                             allTraces.add(entry);
                         }
@@ -2891,113 +2240,8 @@ public class CamelMonitor extends CamelCommand {
         }
     }
 
-    @SuppressWarnings("unchecked")
-    private TraceEntry parseTraceEntry(JsonObject json, String pid) {
-        TraceEntry entry = new TraceEntry();
-        entry.pid = pid;
-        entry.uid = stringValue(json.get("uid"));
-        entry.exchangeId = json.getString("exchangeId");
-        entry.routeId = json.getString("routeId");
-        entry.nodeId = json.getString("nodeId");
-        entry.nodeShortName = json.getString("nodeShortName");
-        entry.location = json.getString("location");
-        entry.nodeLabel = json.getString("nodeLabel");
-        entry.threadName = json.getString("threadName");
-        entry.first = json.getBooleanOrDefault("first", false);
-        entry.last = json.getBooleanOrDefault("last", false);
-        entry.nodeLevel = json.getIntegerOrDefault("nodeLevel", 0);
-
-        // Derive status from done/failed booleans
-        boolean done = Boolean.TRUE.equals(json.get("done"));
-        boolean failed = Boolean.TRUE.equals(json.get("failed"));
-        entry.failed = failed;
-        if (entry.failed) {
-            entry.status = "Failed";
-        } else if (done) {
-            entry.status = "Done";
-        } else {
-            entry.status = "Processing";
-        }
-
-        Object elapsedObj = json.get("elapsed");
-        if (elapsedObj instanceof Number n) {
-            entry.elapsed = n.longValue();
-        } else if (elapsedObj != null) {
-            try {
-                entry.elapsed = Long.parseLong(elapsedObj.toString());
-            } catch (NumberFormatException e) {
-                // ignore
-            }
-        }
-
-        // Timestamp — last entries carry the start time, so add elapsed to get completion time
-        Object tsObj = json.get("timestamp");
-        if (tsObj instanceof Number n) {
-            long epochMs = n.longValue();
-            if (entry.last && entry.elapsed > 0) {
-                epochMs += entry.elapsed;
-            }
-            entry.epochMs = epochMs;
-            entry.timestamp = Instant.ofEpochMilli(epochMs)
-                    .atZone(ZoneId.systemDefault())
-                    .toLocalTime().toString();
-            if (entry.timestamp.length() > 12) {
-                entry.timestamp = entry.timestamp.substring(0, 12);
-            }
-        } else if (tsObj != null) {
-            entry.timestamp = tsObj.toString();
-        }
-
-        // Compute direction and processor label
-        if (entry.first || entry.last) {
-            entry.nodeLevel = Math.max(0, entry.nodeLevel - 1);
-        }
-        String indent = "  ".repeat(entry.nodeLevel);
-        if (entry.first) {
-            entry.direction = "-->";
-            String uri = json.getString("endpointUri");
-            if (uri != null) {
-                entry.processor = indent + "from[" + uri + "]";
-            } else {
-                entry.processor = indent + (entry.nodeLabel != null ? entry.nodeLabel : "");
-            }
-        } else if (entry.last) {
-            entry.direction = "<--";
-            entry.processor = indent + (entry.nodeLabel != null ? entry.nodeLabel : "");
-        } else {
-            entry.direction = "  >";
-            entry.processor = indent + (entry.nodeLabel != null ? entry.nodeLabel : "");
-        }
-
-        // Parse message object
-        Object msgObj = json.get("message");
-        if (msgObj instanceof JsonObject message) {
-            MessageData md = parseMessage(message);
-            entry.headers = md.headers();
-            entry.headerTypes = md.headerTypes();
-            entry.body = md.body();
-            entry.bodyType = md.bodyType();
-            if (entry.body != null) {
-                entry.bodyPreview = entry.body.replace("\n", " ").replace("\r", "");
-            }
-            entry.exchangeProperties = md.exchangeProperties();
-            entry.exchangePropertyTypes = md.exchangePropertyTypes();
-            entry.exchangeVariables = md.exchangeVariables();
-            entry.exchangeVariableTypes = md.exchangeVariableTypes();
-        }
-
-        // Exception (message + full stacktrace)
-        Object excObj = json.get("exception");
-        if (excObj instanceof JsonObject excJson) {
-            String msg = excJson.getString("message");
-            entry.exception = msg != null ? Jsoner.unescape(msg) : null;
-            String st = excJson.getString("stackTrace");
-            if (st != null && !st.isEmpty()) {
-                entry.exception = entry.exception + "\n" + Jsoner.unescape(st);
-            }
-        }
-
-        return entry;
+    private JsonObject loadErrorFile(long pid) {
+        return TuiHelper.loadStatus(pid, this::getErrorFile);
     }
 
     @SuppressWarnings("unchecked")
@@ -3018,7 +2262,7 @@ public class CamelMonitor extends CamelCommand {
                 if (tracesArray instanceof List<?> traceList) {
                     for (Object traceObj : traceList) {
                         if (traceObj instanceof JsonObject traceJson) {
-                            HistoryEntry entry = parseHistoryEntry(traceJson, Long.toString(pid));
+                            HistoryEntry entry = StatusParser.parseHistoryEntry(traceJson, Long.toString(pid));
                             if (entry != null) {
                                 allEntries.add(entry);
                             }
@@ -3030,839 +2274,6 @@ public class CamelMonitor extends CamelCommand {
             }
         }
         historyTab.historyEntries = allEntries;
-    }
-
-    @SuppressWarnings("unchecked")
-    private HistoryEntry parseHistoryEntry(JsonObject json, String pid) {
-        HistoryEntry entry = new HistoryEntry();
-        entry.pid = pid;
-        entry.exchangeId = json.getString("exchangeId");
-        entry.routeId = json.getString("routeId");
-        entry.fromRouteId = json.getString("fromRouteId");
-        entry.nodeId = json.getString("nodeId");
-        entry.nodeShortName = json.getString("nodeShortName");
-        entry.nodeLabel = json.getString("nodeLabel");
-        entry.location = json.getString("location");
-        entry.threadName = json.getString("threadName");
-        entry.first = json.getBooleanOrDefault("first", false);
-        entry.last = json.getBooleanOrDefault("last", false);
-        entry.failed = json.getBooleanOrDefault("failed", false);
-        entry.nodeLevel = json.getIntegerOrDefault("nodeLevel", 0);
-
-        Object elapsedObj = json.get("elapsed");
-        if (elapsedObj instanceof Number n) {
-            entry.elapsed = n.longValue();
-        } else {
-            entry.elapsed = -1;
-        }
-
-        // Compute direction arrow
-        if (entry.first) {
-            entry.direction = "-->";
-        } else if (entry.last) {
-            entry.direction = "<--";
-        } else {
-            entry.direction = "  >";
-        }
-
-        // Compute processor label with tree indentation
-        if (entry.first || entry.last) {
-            entry.nodeLevel = Math.max(0, entry.nodeLevel - 1);
-        }
-        String indent = "  ".repeat(entry.nodeLevel);
-        if (entry.first) {
-            String uri = json.getString("endpointUri");
-            if (uri != null) {
-                entry.processor = indent + "from[" + uri + "]";
-            } else {
-                entry.processor = indent + (entry.nodeLabel != null ? entry.nodeLabel : "");
-            }
-        } else {
-            entry.processor = indent + (entry.nodeLabel != null ? entry.nodeLabel : "");
-        }
-
-        // Timestamp — last entries carry the start time, so add elapsed to get completion time
-        Object tsObj = json.get("timestamp");
-        if (tsObj instanceof Number n) {
-            long epochMs = n.longValue();
-            if (entry.last && entry.elapsed > 0) {
-                epochMs += entry.elapsed;
-            }
-            entry.epochMs = epochMs;
-            entry.timestamp = Instant.ofEpochMilli(epochMs)
-                    .atZone(ZoneId.systemDefault())
-                    .toLocalTime().toString();
-            if (entry.timestamp.length() > 12) {
-                entry.timestamp = entry.timestamp.substring(0, 12);
-            }
-        }
-
-        // Parse message
-        Object msgObj = json.get("message");
-        if (msgObj instanceof JsonObject message) {
-            MessageData md = parseMessage(message);
-            entry.headers = md.headers();
-            entry.headerTypes = md.headerTypes();
-            entry.body = md.body();
-            entry.bodyType = md.bodyType();
-            entry.exchangeProperties = md.exchangeProperties();
-            entry.exchangePropertyTypes = md.exchangePropertyTypes();
-            entry.exchangeVariables = md.exchangeVariables();
-            entry.exchangeVariableTypes = md.exchangeVariableTypes();
-        }
-
-        // Exception (message + full stacktrace)
-        Object excObj = json.get("exception");
-        if (excObj instanceof JsonObject excJson) {
-            String msg = excJson.getString("message");
-            entry.exception = msg != null ? Jsoner.unescape(msg) : null;
-            String st = excJson.getString("stackTrace");
-            if (st != null && !st.isEmpty()) {
-                entry.exception = entry.exception + "\n" + Jsoner.unescape(st);
-            }
-        }
-
-        return entry;
-    }
-
-    private static String stringValue(Object obj) {
-        return obj != null ? obj.toString() : null;
-    }
-
-    record MessageData(
-            Map<String, Object> headers,
-            Map<String, String> headerTypes,
-            String body,
-            String bodyType,
-            Map<String, Object> exchangeProperties,
-            Map<String, String> exchangePropertyTypes,
-            Map<String, Object> exchangeVariables,
-            Map<String, String> exchangeVariableTypes) {
-    }
-
-    @SuppressWarnings("unchecked")
-    private static MessageData parseMessage(JsonObject message) {
-        Map<String, Object> headers = null;
-        Map<String, String> headerTypes = null;
-        String body = null;
-        String bodyType = null;
-        Map<String, Object> exchangeProperties = null;
-        Map<String, String> exchangePropertyTypes = null;
-        Map<String, Object> exchangeVariables = null;
-        Map<String, String> exchangeVariableTypes = null;
-
-        // Headers
-        Object headersObj = message.get("headers");
-        if (headersObj instanceof List<?> headerList) {
-            headers = new LinkedHashMap<>();
-            headerTypes = new LinkedHashMap<>();
-            for (Object h : headerList) {
-                if (h instanceof JsonObject hObj) {
-                    String key = String.valueOf(hObj.get("key"));
-                    headers.put(key, hObj.get("value"));
-                    Object type = hObj.get("type");
-                    if (type != null) {
-                        headerTypes.put(key, TuiHelper.shortTypeName(type.toString()));
-                    }
-                }
-            }
-        } else if (headersObj instanceof Map) {
-            headers = new LinkedHashMap<>((Map<String, Object>) headersObj);
-        }
-
-        // Body
-        Object bodyObj = message.get("body");
-        if (bodyObj instanceof JsonObject bodyJson) {
-            Object val = bodyJson.get("value");
-            body = val != null ? val.toString() : null;
-            bodyType = TuiHelper.shortTypeName(bodyJson.getString("type"));
-        } else if (bodyObj != null) {
-            body = bodyObj.toString();
-        }
-
-        // Exchange properties
-        Object propsObj = message.get("exchangeProperties");
-        if (propsObj instanceof List<?> propList) {
-            exchangeProperties = new LinkedHashMap<>();
-            exchangePropertyTypes = new LinkedHashMap<>();
-            for (Object p : propList) {
-                if (p instanceof JsonObject pObj) {
-                    String key = String.valueOf(pObj.get("key"));
-                    exchangeProperties.put(key, pObj.get("value"));
-                    Object type = pObj.get("type");
-                    if (type != null) {
-                        exchangePropertyTypes.put(key, TuiHelper.shortTypeName(type.toString()));
-                    }
-                }
-            }
-        } else if (propsObj instanceof Map) {
-            exchangeProperties = new LinkedHashMap<>((Map<String, Object>) propsObj);
-        }
-
-        // Exchange variables
-        Object varsObj = message.get("exchangeVariables");
-        if (varsObj instanceof List<?> varList) {
-            exchangeVariables = new LinkedHashMap<>();
-            exchangeVariableTypes = new LinkedHashMap<>();
-            for (Object v : varList) {
-                if (v instanceof JsonObject vObj) {
-                    String key = String.valueOf(vObj.get("key"));
-                    exchangeVariables.put(key, vObj.get("value"));
-                    Object type = vObj.get("type");
-                    if (type != null) {
-                        exchangeVariableTypes.put(key, TuiHelper.shortTypeName(type.toString()));
-                    }
-                }
-            }
-        } else if (varsObj instanceof Map) {
-            exchangeVariables = new LinkedHashMap<>((Map<String, Object>) varsObj);
-        }
-
-        return new MessageData(
-                headers, headerTypes, body, bodyType,
-                exchangeProperties, exchangePropertyTypes, exchangeVariables, exchangeVariableTypes);
-    }
-
-    // ---- Integration Parsing ----
-
-    @SuppressWarnings("unchecked")
-    private IntegrationInfo parseIntegration(ProcessHandle ph, JsonObject root) {
-        JsonObject context = (JsonObject) root.get("context");
-        if (context == null) {
-            return null;
-        }
-
-        IntegrationInfo info = new IntegrationInfo();
-        info.name = context.getString("name");
-        if ("CamelJBang".equals(info.name)) {
-            info.name = ProcessHelper.extractName(root, ph);
-        }
-        info.pid = Long.toString(ph.pid());
-        info.uptime = extractSince(ph);
-        info.ago = TimeUtils.printSince(info.uptime);
-        info.state = context.getIntegerOrDefault("phase", 0);
-        info.camelVersion = context.getString("version");
-        info.profile = context.getString("profile");
-        info.devMode = context.getBooleanOrDefault("devMode", false);
-
-        JsonObject runtime = (JsonObject) root.get("runtime");
-        info.platform = runtime != null ? runtime.getString("platform") : null;
-        info.platformVersion = runtime != null ? runtime.getString("platformVersion") : null;
-        if ("Camel".equals(info.platform)) {
-            String cl = ph.info().commandLine().orElse("");
-            if (cl.contains("main.CamelJBang run")) {
-                info.platform = "JBang";
-                if (info.platformVersion == null) {
-                    info.platformVersion = VersionHelper.getJBangVersion();
-                }
-            }
-        }
-        info.directory = runtime != null ? runtime.getString("directory") : null;
-        info.javaVersion = runtime != null ? runtime.getString("javaVersion") : null;
-        info.javaVendor = runtime != null ? runtime.getString("javaVendor") : null;
-        info.javaVmName = runtime != null ? runtime.getString("javaVmName") : null;
-        info.readmeFiles = runtime != null ? runtime.getString("readmeFiles") : null;
-
-        Map<String, ?> stats = context.getMap("statistics");
-        if (stats != null) {
-            Object thp = stats.get("exchangesThroughput");
-            if (thp != null) {
-                info.throughput = thp.toString();
-            }
-            info.exchangesTotal = objToLong(stats.get("exchangesTotal"));
-            info.failed = objToLong(stats.get("exchangesFailed"));
-            info.inflight = objToLong(stats.get("exchangesInflight"));
-            info.inflightLoad01 = objToString(stats.get("load01"));
-            info.inflightLoad05 = objToString(stats.get("load05"));
-            info.inflightLoad15 = objToString(stats.get("load15"));
-            info.last = objToString(stats.get("lastProcessingTime"));
-            info.delta = objToString(stats.get("deltaProcessingTime"));
-            long tsStarted = objToLong(stats.get("lastCreatedExchangeTimestamp"));
-            if (tsStarted > 0) {
-                info.sinceLastStarted = TimeUtils.printSince(tsStarted);
-            }
-            long tsCompleted = objToLong(stats.get("lastCompletedExchangeTimestamp"));
-            if (tsCompleted > 0) {
-                info.sinceLastCompleted = TimeUtils.printSince(tsCompleted);
-            }
-            long tsFailed = objToLong(stats.get("lastFailedExchangeTimestamp"));
-            if (tsFailed > 0) {
-                info.sinceLastFailed = TimeUtils.printSince(tsFailed);
-            }
-            Map<String, ?> reloadStats = (Map<String, ?>) stats.get("reload");
-            if (reloadStats != null) {
-                info.reloaded = (int) objToLong(reloadStats.get("reloaded"));
-            }
-        }
-
-        JsonObject mem = (JsonObject) root.get("memory");
-        if (mem != null) {
-            info.heapMemUsed = mem.getLongOrDefault("heapMemoryUsed", 0L);
-            info.heapMemCommitted = mem.getLongOrDefault("heapMemoryCommitted", 0L);
-            info.heapMemMax = mem.getLongOrDefault("heapMemoryMax", 0L);
-            info.nonHeapMemUsed = mem.getLongOrDefault("nonHeapMemoryUsed", 0L);
-            info.nonHeapMemCommitted = mem.getLongOrDefault("nonHeapMemoryCommitted", 0L);
-            info.oldGenUsed = mem.getLongOrDefault("oldGenUsed", 0L);
-            info.oldGenCommitted = mem.getLongOrDefault("oldGenCommitted", 0L);
-            info.oldGenMax = mem.getLongOrDefault("oldGenMax", 0L);
-            info.metaspaceUsed = mem.getLongOrDefault("metaspaceUsed", 0L);
-            info.metaspaceCommitted = mem.getLongOrDefault("metaspaceCommitted", 0L);
-            info.metaspaceMax = mem.getLongOrDefault("metaspaceMax", 0L);
-        }
-
-        JsonObject gc = (JsonObject) root.get("gc");
-        if (gc != null) {
-            info.gcCollectionCount = gc.getLongOrDefault("collectionCount", 0L);
-            info.gcCollectionTime = gc.getLongOrDefault("collectionTime", 0L);
-        }
-
-        JsonObject classLoading = (JsonObject) root.get("classLoading");
-        if (classLoading != null) {
-            info.loadedClassCount = classLoading.getIntegerOrDefault("loadedClassCount", 0);
-            info.totalLoadedClassCount = classLoading.getLongOrDefault("totalLoadedClassCount", 0L);
-        }
-
-        JsonObject threads = (JsonObject) root.get("threads");
-        if (threads != null) {
-            info.threadCount = threads.getIntegerOrDefault("threadCount", 0);
-            info.peakThreadCount = threads.getIntegerOrDefault("peakThreadCount", 0);
-        }
-
-        JsonObject logger = (JsonObject) root.get("logger");
-        if (logger != null) {
-            JsonObject levels = (JsonObject) logger.get("levels");
-            if (levels != null) {
-                info.rootLogLevel = levels.getString("root");
-            }
-        }
-
-        // Parse routes
-        JsonArray routes = (JsonArray) root.get("routes");
-        if (routes != null) {
-            for (Object r : routes) {
-                JsonObject rj = (JsonObject) r;
-                RouteInfo ri = new RouteInfo();
-                ri.routeId = rj.getString("routeId");
-                ri.group = rj.getString("group");
-                ri.from = rj.getString("from");
-                ri.state = rj.getString("state");
-                ri.supportsSuspension = rj.getBooleanOrDefault("supportsSuspension", false);
-                ri.uptime = rj.getString("uptime");
-
-                Map<String, ?> rs = rj.getMap("statistics");
-                if (rs != null) {
-                    ri.coverage = objToString(rs.get("coverage"));
-                    ri.throughput = objToString(rs.get("exchangesThroughput"));
-                    ri.total = objToLong(rs.get("exchangesTotal"));
-                    ri.failed = objToLong(rs.get("exchangesFailed"));
-                    ri.inflight = objToLong(rs.get("exchangesInflight"));
-                    ri.meanTime = Math.max(0, objToLong(rs.get("meanProcessingTime")));
-                    ri.minTime = Math.max(0, objToLong(rs.get("minProcessingTime")));
-                    ri.maxTime = Math.max(0, objToLong(rs.get("maxProcessingTime")));
-                    ri.lastTime = Math.max(0, objToLong(rs.get("lastProcessingTime")));
-                    ri.deltaTime = objToLong(rs.get("deltaProcessingTime"));
-                    ri.load01 = objToString(rs.get("load01"));
-                    ri.load05 = objToString(rs.get("load05"));
-                    ri.load15 = objToString(rs.get("load15"));
-                    long tsStarted = objToLong(rs.get("lastCreatedExchangeTimestamp"));
-                    if (tsStarted > 0) {
-                        ri.sinceLastStarted = TimeUtils.printSince(tsStarted);
-                    }
-                    long tsCompleted = objToLong(rs.get("lastCompletedExchangeTimestamp"));
-                    if (tsCompleted > 0) {
-                        ri.sinceLastCompleted = TimeUtils.printSince(tsCompleted);
-                    }
-                    long tsFailed = objToLong(rs.get("lastFailedExchangeTimestamp"));
-                    if (tsFailed > 0) {
-                        ri.sinceLastFailed = TimeUtils.printSince(tsFailed);
-                    }
-                }
-
-                // Parse processors
-                JsonArray procs = (JsonArray) rj.get("processors");
-                if (procs != null) {
-                    for (Object p : procs) {
-                        JsonObject pj = (JsonObject) p;
-                        ProcessorInfo pi = new ProcessorInfo();
-                        pi.id = pj.getString("id");
-                        pi.processor = pj.getString("processor");
-                        pi.level = pj.getIntegerOrDefault("level", 0);
-
-                        Map<String, ?> ps = pj.getMap("statistics");
-                        if (ps != null) {
-                            pi.total = objToLong(ps.get("exchangesTotal"));
-                            pi.failed = objToLong(ps.get("exchangesFailed"));
-                            pi.meanTime = Math.max(0, objToLong(ps.get("meanProcessingTime")));
-                            pi.minTime = Math.max(0, objToLong(ps.get("minProcessingTime")));
-                            pi.maxTime = Math.max(0, objToLong(ps.get("maxProcessingTime")));
-                            pi.lastTime = objToLong(ps.get("lastProcessingTime"));
-                            pi.deltaTime = objToLong(ps.get("deltaProcessingTime"));
-                            pi.inflight = objToLong(ps.get("exchangesInflight"));
-                            long tsStarted = objToLong(ps.get("lastCreatedExchangeTimestamp"));
-                            if (tsStarted > 0) {
-                                pi.sinceLastStarted = TimeUtils.printSince(tsStarted);
-                            }
-                            long tsCompleted = objToLong(ps.get("lastCompletedExchangeTimestamp"));
-                            if (tsCompleted > 0) {
-                                pi.sinceLastCompleted = TimeUtils.printSince(tsCompleted);
-                            }
-                            long tsFailed = objToLong(ps.get("lastFailedExchangeTimestamp"));
-                            if (tsFailed > 0) {
-                                pi.sinceLastFailed = TimeUtils.printSince(tsFailed);
-                            }
-                        }
-
-                        ri.processors.add(pi);
-                    }
-                }
-
-                info.routes.add(ri);
-            }
-            info.routeTotal = info.routes.size();
-            info.routeStarted = (int) info.routes.stream().filter(r -> "Started".equals(r.state)).count();
-        }
-
-        // Parse health checks and ready status
-        JsonObject healthChecks = (JsonObject) root.get("healthChecks");
-        if (healthChecks != null) {
-            Boolean rdy = (Boolean) healthChecks.get("ready");
-            info.ready = Boolean.TRUE.equals(rdy) ? "1/1" : "0/1";
-            JsonArray checks = (JsonArray) healthChecks.get("checks");
-            if (checks != null) {
-                for (Object c : checks) {
-                    JsonObject cj = (JsonObject) c;
-                    HealthCheckInfo hc = new HealthCheckInfo();
-                    hc.group = cj.getString("group");
-                    hc.name = cj.getString("id");
-                    hc.state = cj.getString("state");
-                    hc.readiness = cj.getBooleanOrDefault("readiness", false);
-                    hc.liveness = cj.getBooleanOrDefault("liveness", false);
-                    hc.message = cj.getString("message");
-                    if (hc.message == null) {
-                        JsonObject details = (JsonObject) cj.get("details");
-                        if (details != null && details.containsKey("failure.error.message")) {
-                            hc.message = details.getString("failure.error.message");
-                        }
-                    }
-                    info.healthChecks.add(hc);
-                }
-            }
-        }
-
-        // Parse consumers
-        JsonObject consumersObj = (JsonObject) root.get("consumers");
-        if (consumersObj != null) {
-            JsonArray consumerList = (JsonArray) consumersObj.get("consumers");
-            if (consumerList != null) {
-                for (Object c : consumerList) {
-                    JsonObject cj = (JsonObject) c;
-                    ConsumerInfo ci = new ConsumerInfo();
-                    ci.id = cj.getString("id");
-                    ci.uri = cj.getString("uri");
-                    ci.state = cj.getString("state");
-                    ci.className = cj.getString("class");
-                    ci.scheduled = Boolean.TRUE.equals(cj.get("scheduled"));
-                    ci.inflight = cj.getIntegerOrDefault("inflight", 0);
-                    ci.polling = Boolean.TRUE.equals(cj.get("polling"));
-                    ci.totalCounter = cj.getLong("totalCounter");
-                    ci.delay = cj.getLong("delay");
-                    ci.period = cj.getLong("period");
-                    JsonObject cStats = (JsonObject) cj.get("statistics");
-                    if (cStats != null) {
-                        Object last = cStats.get("lastCreatedExchangeTimestamp");
-                        if (last != null) {
-                            ci.sinceLastStarted = TimeUtils.printSince(Long.parseLong(last.toString()));
-                        }
-                        last = cStats.get("lastCompletedExchangeTimestamp");
-                        if (last != null) {
-                            ci.sinceLastCompleted = TimeUtils.printSince(Long.parseLong(last.toString()));
-                        }
-                        last = cStats.get("lastFailedExchangeTimestamp");
-                        if (last != null) {
-                            ci.sinceLastFailed = TimeUtils.printSince(Long.parseLong(last.toString()));
-                        }
-                    }
-                    info.consumers.add(ci);
-                }
-            }
-        }
-
-        // Parse endpoints (top-level "endpoints" is a JsonObject with nested "endpoints" array)
-        JsonObject endpointsObj = (JsonObject) root.get("endpoints");
-        if (endpointsObj != null) {
-            JsonArray endpointList = (JsonArray) endpointsObj.get("endpoints");
-            if (endpointList != null) {
-                for (Object e : endpointList) {
-                    JsonObject ej = (JsonObject) e;
-                    EndpointInfo ep = new EndpointInfo();
-                    ep.uri = ej.getString("uri");
-                    ep.direction = ej.getString("direction");
-                    ep.routeId = ej.getString("routeId");
-                    ep.hits = TuiHelper.objToLong(ej.get("hits"));
-                    ep.stub = Boolean.TRUE.equals(ej.get("stub"));
-                    ep.remote = !Boolean.FALSE.equals(ej.get("remote"));
-                    ep.minBodySize = TuiHelper.objToLong(ej.get("minBodySize"));
-                    ep.maxBodySize = TuiHelper.objToLong(ej.get("maxBodySize"));
-                    ep.meanBodySize = TuiHelper.objToLong(ej.get("meanBodySize"));
-                    ep.minHeadersSize = TuiHelper.objToLong(ej.get("minHeadersSize"));
-                    ep.maxHeadersSize = TuiHelper.objToLong(ej.get("maxHeadersSize"));
-                    ep.meanHeadersSize = TuiHelper.objToLong(ej.get("meanHeadersSize"));
-                    // Extract component from URI (e.g., "timer://tick" -> "timer")
-                    if (ep.uri != null) {
-                        int idx = ep.uri.indexOf(':');
-                        ep.component = idx > 0 ? ep.uri.substring(0, idx) : ep.uri;
-                    }
-                    info.endpoints.add(ep);
-                }
-            }
-        }
-
-        // Parse circuit breakers: resilience4j, fault-tolerance, core
-        parseCbSection(root, "resilience4j", info);
-        parseCbSection(root, "fault-tolerance", info);
-        parseCbSection(root, "circuit-breaker", info);
-
-        // Enrich circuit breakers with processor statistics (matched by id)
-        for (CircuitBreakerInfo cb : info.circuitBreakers) {
-            if (cb.id != null) {
-                for (RouteInfo ri : info.routes) {
-                    for (ProcessorInfo pi : ri.processors) {
-                        if (cb.id.equals(pi.id)) {
-                            cb.total = pi.total;
-                            cb.totalFailed = pi.failed;
-                            cb.meanTime = pi.meanTime;
-                            cb.minTime = pi.minTime;
-                            cb.maxTime = pi.maxTime;
-                            cb.inflight = pi.inflight;
-                            cb.sinceLastStarted = pi.sinceLastStarted;
-                            cb.sinceLastSuccess = pi.sinceLastCompleted;
-                            cb.sinceLastFail = pi.sinceLastFailed;
-                            break;
-                        }
-                    }
-                }
-            }
-        }
-
-        // Parse error count from error registry (full error data is loaded on demand by ErrorsTab)
-        JsonObject errorsObj = (JsonObject) root.get("errors");
-        if (errorsObj != null) {
-            info.errorCount = errorsObj.getIntegerOrDefault("size", 0);
-        }
-
-        // Parse inflight exchanges
-        JsonObject inflightObj = (JsonObject) root.get("inflight");
-        if (inflightObj != null) {
-            info.inflightBrowseEnabled = inflightObj.getBooleanOrDefault("inflightBrowseEnabled", false);
-            int inflightCount = inflightObj.getIntegerOrDefault("inflight", 0);
-            if (inflightCount > 0) {
-                JsonArray inflArr = (JsonArray) inflightObj.get("exchanges");
-                if (inflArr != null) {
-                    for (Object ie : inflArr) {
-                        JsonObject ij = (JsonObject) ie;
-                        InflightInfo ii = new InflightInfo();
-                        ii.exchangeId = ij.getString("exchangeId");
-                        ii.fromRouteId = ij.getString("fromRouteId");
-                        Boolean remote = ij.getBoolean("fromRemoteEndpoint");
-                        ii.fromRemoteEndpoint = remote != null && remote;
-                        ii.atRouteId = ij.getString("atRouteId");
-                        ii.nodeId = ij.getString("nodeId");
-                        ii.elapsed = ij.getLongOrDefault("elapsed", 0L);
-                        ii.duration = ij.getLongOrDefault("duration", 0L);
-                        ii.blocked = false;
-                        info.inflightExchanges.add(ii);
-                    }
-                }
-            }
-        }
-
-        // Parse blocked exchanges
-        JsonObject blockedObj = (JsonObject) root.get("blocked");
-        if (blockedObj != null) {
-            int blockedCount = blockedObj.getIntegerOrDefault("blocked", 0);
-            if (blockedCount > 0) {
-                JsonArray blkArr = (JsonArray) blockedObj.get("exchanges");
-                if (blkArr != null) {
-                    for (Object be : blkArr) {
-                        JsonObject bj = (JsonObject) be;
-                        InflightInfo ii = new InflightInfo();
-                        ii.exchangeId = bj.getString("exchangeId");
-                        ii.atRouteId = bj.getString("routeId");
-                        ii.nodeId = bj.getString("nodeId");
-                        ii.duration = bj.getLongOrDefault("duration", 0L);
-                        ii.blocked = true;
-                        info.inflightExchanges.add(ii);
-                    }
-                }
-            }
-        }
-
-        // Parse micrometer metrics (optional, only present when --observe is used)
-        JsonObject micrometerObj = (JsonObject) root.get("micrometer");
-        if (micrometerObj != null) {
-            parseMicrometerMeters(micrometerObj, "counters", "counter", info);
-            parseMicrometerMeters(micrometerObj, "gauges", "gauge", info);
-            parseMicrometerMeters(micrometerObj, "timers", "timer", info);
-            parseMicrometerMeters(micrometerObj, "longTaskTimers", "longTaskTimer", info);
-            parseMicrometerMeters(micrometerObj, "distribution", "distribution", info);
-        }
-
-        // Parse REST DSL services
-        JsonObject restsObj = (JsonObject) root.get("rests");
-        if (restsObj != null) {
-            JsonArray restList = (JsonArray) restsObj.get("rests");
-            if (restList != null) {
-                for (Object r : restList) {
-                    JsonObject rj = (JsonObject) r;
-                    HttpEndpointInfo ep = new HttpEndpointInfo();
-                    ep.fromRest = true;
-                    ep.url = rj.getString("url");
-                    ep.method = rj.getString("method");
-                    if (ep.method != null) {
-                        ep.method = ep.method.toUpperCase(Locale.ENGLISH);
-                    }
-                    ep.consumes = rj.getString("consumes");
-                    ep.produces = rj.getString("produces");
-                    ep.description = rj.getString("description");
-                    ep.contractFirst = Boolean.TRUE.equals(rj.get("contractFirst"));
-                    ep.specification = Boolean.TRUE.equals(rj.get("specification"));
-                    ep.routeId = rj.getString("routeId");
-                    ep.operationId = rj.getString("operationId");
-                    ep.specificationUri = rj.getString("specificationUri");
-                    ep.state = rj.getString("state");
-                    ep.inType = rj.getString("inType");
-                    ep.outType = rj.getString("outType");
-                    Long h = rj.getLong("hits");
-                    if (h != null) {
-                        ep.hits = h;
-                    }
-                    // derive path from url (strip scheme+host+port)
-                    ep.path = extractPath(ep.url);
-                    info.httpEndpoints.add(ep);
-                }
-            }
-        }
-
-        // Parse Platform-HTTP services
-        JsonObject phpObj = (JsonObject) root.get("platform-http");
-        if (phpObj != null) {
-            info.httpServer = phpObj.getString("server");
-            parseHttpEndpoints(phpObj, "endpoints", false, info);
-            parseHttpEndpoints(phpObj, "managementEndpoints", true, info);
-        }
-
-        // Fix REST DSL URLs that are missing the port by resolving from platform-http endpoints
-        fixRestUrlPorts(info);
-
-        // Parse configuration properties (from PropertiesDevConsole)
-        JsonObject propsObj = (JsonObject) root.get("properties");
-        if (propsObj != null) {
-            JsonArray propArr = (JsonArray) propsObj.get("properties");
-            if (propArr != null) {
-                for (Object p : propArr) {
-                    JsonObject pj = (JsonObject) p;
-                    String key = pj.getString("key");
-                    if (key != null && !key.startsWith("camel.jbang.")) {
-                        ConfigurationTab.ConfigProperty cp = new ConfigurationTab.ConfigProperty();
-                        cp.key = key;
-                        cp.value = objToString(pj.get("value"));
-                        cp.defaultValue = pj.getString("defaultValue");
-                        cp.source = pj.getString("source");
-                        cp.location = pj.getString("location");
-                        info.configProperties.add(cp);
-                    }
-                }
-                info.configProperties.sort(ConfigurationTab::compareCamelFirst);
-            }
-        }
-
-        return info;
-    }
-
-    private static void parseHttpEndpoints(JsonObject phpObj, String key, boolean management, IntegrationInfo info) {
-        JsonArray arr = (JsonArray) phpObj.get(key);
-        if (arr == null) {
-            return;
-        }
-        for (Object e : arr) {
-            JsonObject ej = (JsonObject) e;
-            HttpEndpointInfo ep = new HttpEndpointInfo();
-            ep.fromRest = false;
-            ep.management = management;
-            ep.server = phpObj.getString("server");
-            ep.url = ej.getString("url");
-            ep.path = ej.getString("path");
-            ep.method = ej.getString("verbs");
-            ep.consumes = ej.getString("consumes");
-            ep.produces = ej.getString("produces");
-            info.httpEndpoints.add(ep);
-        }
-    }
-
-    private static void fixRestUrlPorts(IntegrationInfo info) {
-        // Find a base URL with port from platform-http endpoints
-        String baseWithPort = null;
-        for (HttpEndpointInfo ep : info.httpEndpoints) {
-            if (!ep.fromRest && ep.url != null) {
-                try {
-                    URI uri = URI.create(ep.url);
-                    if (uri.getPort() > 0) {
-                        baseWithPort = uri.getScheme() + "://" + uri.getHost() + ":" + uri.getPort();
-                        break;
-                    }
-                } catch (Exception e) {
-                    // ignore
-                }
-            }
-        }
-        if (baseWithPort == null) {
-            return;
-        }
-        // Fix REST DSL endpoints missing port
-        for (HttpEndpointInfo ep : info.httpEndpoints) {
-            if (ep.fromRest && ep.url != null) {
-                String path = extractPath(ep.url);
-                if (path != null) {
-                    // Check if URL has no port by looking at the host portion
-                    int schemeEnd = ep.url.indexOf("://");
-                    if (schemeEnd > 0) {
-                        String hostPart = ep.url.substring(schemeEnd + 3);
-                        int slash = hostPart.indexOf('/');
-                        if (slash > 0) {
-                            hostPart = hostPart.substring(0, slash);
-                        }
-                        if (!hostPart.contains(":")) {
-                            ep.url = baseWithPort + path;
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    private static String extractPath(String url) {
-        if (url == null) {
-            return null;
-        }
-        // strip scheme://host:port prefix — find third '/' or return url as-is
-        int idx = url.indexOf("://");
-        if (idx < 0) {
-            return url;
-        }
-        int slash = url.indexOf('/', idx + 3);
-        return slash >= 0 ? url.substring(slash) : "/";
-    }
-
-    private static void parseCbSection(JsonObject root, String key, IntegrationInfo info) {
-        JsonObject section = (JsonObject) root.get(key);
-        if (section == null) {
-            return;
-        }
-        JsonArray breakers = (JsonArray) section.get("circuitBreakers");
-        if (breakers == null) {
-            return;
-        }
-        String component = switch (key) {
-            case "resilience4j" -> "resilience4j";
-            case "fault-tolerance" -> "fault-tolerance";
-            default -> "core";
-        };
-        for (Object b : breakers) {
-            JsonObject bj = (JsonObject) b;
-            CircuitBreakerInfo cb = new CircuitBreakerInfo();
-            cb.component = component;
-            cb.routeId = bj.getString("routeId");
-            cb.id = bj.getString("id");
-            cb.state = bj.getString("state");
-            cb.bufferedCalls = bj.getIntegerOrDefault("bufferedCalls", 0);
-            cb.successfulCalls = TuiHelper.objToLong(bj.get("successfulCalls"));
-            cb.failedCalls = TuiHelper.objToLong(bj.get("failedCalls"));
-            cb.notPermittedCalls = TuiHelper.objToLong(bj.get("notPermittedCalls"));
-            Object fr = bj.get("failureRate");
-            cb.failureRate = fr instanceof Number n ? n.doubleValue() : -1;
-            info.circuitBreakers.add(cb);
-        }
-    }
-
-    @SuppressWarnings("unchecked")
-    private static void parseMicrometerMeters(
-            JsonObject micrometerObj, String section, String type, IntegrationInfo info) {
-        JsonArray arr = (JsonArray) micrometerObj.get(section);
-        if (arr == null) {
-            return;
-        }
-        for (Object o : arr) {
-            JsonObject jo = (JsonObject) o;
-            MicrometerMeterInfo m = new MicrometerMeterInfo();
-            m.type = type;
-            m.name = jo.getString("name");
-            m.description = jo.getString("description");
-            // parse tags
-            JsonArray tagsArr = (JsonArray) jo.get("tags");
-            if (tagsArr != null) {
-                for (Object t : tagsArr) {
-                    JsonObject tj = (JsonObject) t;
-                    m.tags.add(new String[] { tj.getString("key"), tj.getString("value") });
-                }
-            }
-            // parse type-specific values
-            switch (type) {
-                case "counter":
-                    m.count = TuiHelper.objToLong(jo.get("count"));
-                    break;
-                case "gauge":
-                    Object v = jo.get("value");
-                    m.value = v instanceof Number n ? n.doubleValue() : null;
-                    break;
-                case "timer":
-                    m.count = TuiHelper.objToLong(jo.get("count"));
-                    m.mean = TuiHelper.objToLong(jo.get("mean"));
-                    m.max = TuiHelper.objToLong(jo.get("max"));
-                    m.total = TuiHelper.objToLong(jo.get("total"));
-                    break;
-                case "longTaskTimer":
-                    Object at = jo.get("activeTasks");
-                    m.activeTasks = at instanceof Number n ? n.intValue() : null;
-                    m.mean = TuiHelper.objToLong(jo.get("mean"));
-                    m.max = TuiHelper.objToLong(jo.get("max"));
-                    m.total = TuiHelper.objToLong(jo.get("duration"));
-                    break;
-                case "distribution":
-                    m.count = TuiHelper.objToLong(jo.get("count"));
-                    Object dm = jo.get("mean");
-                    m.meanDouble = dm instanceof Number n ? n.doubleValue() : null;
-                    Object dx = jo.get("max");
-                    m.maxDouble = dx instanceof Number n ? n.doubleValue() : null;
-                    Object dt = jo.get("totalAmount");
-                    m.totalDouble = dt instanceof Number n ? n.doubleValue() : null;
-                    break;
-                default:
-                    break;
-            }
-            info.meters.add(m);
-        }
-    }
-
-    @SuppressWarnings("unchecked")
-    private static void parseKvArray(JsonArray arr, Map<String, Object> values, Map<String, String> types) {
-        if (arr == null) {
-            return;
-        }
-        for (Object o : arr) {
-            JsonObject jo = (JsonObject) o;
-            String key = jo.getString("key");
-            if (key != null) {
-                values.put(key, jo.get("value"));
-                String type = jo.getString("type");
-                if (type != null) {
-                    types.put(key, type);
-                }
-            }
-        }
-    }
-
-    private JsonObject loadErrorFile(long pid) {
-        return TuiHelper.loadStatus(pid, this::getErrorFile);
     }
 
     private void refreshErrorData(List<Long> pids) {
@@ -3927,17 +2338,17 @@ public class CamelMonitor extends CamelCommand {
                     }
                     JsonArray hdrs = msg.getCollection("headers");
                     if (hdrs != null) {
-                        parseKvArray(hdrs, ei.headers, ei.headerTypes);
+                        StatusParser.parseKvArray(hdrs, ei.headers, ei.headerTypes);
                     }
                 }
                 // exchange properties and variables
                 JsonArray props = ej.getCollection("exchangeProperties");
                 if (props != null) {
-                    parseKvArray(props, ei.properties, ei.propertyTypes);
+                    StatusParser.parseKvArray(props, ei.properties, ei.propertyTypes);
                 }
                 JsonArray vars = ej.getCollection("exchangeVariables");
                 if (vars != null) {
-                    parseKvArray(vars, ei.variables, ei.variableTypes);
+                    StatusParser.parseKvArray(vars, ei.variables, ei.variableTypes);
                 }
                 parsed.add(ei);
             }
@@ -3974,10 +2385,6 @@ public class CamelMonitor extends CamelCommand {
         return TuiHelper.loadStatus(pid, this::getStatusFile);
     }
 
-    private static long extractSince(ProcessHandle ph) {
-        return ph.info().startInstant().map(Instant::toEpochMilli).orElse(0L);
-    }
-
     private void takeScreenshot() {
         Buffer buf = lastBuffer;
         if (buf == null) {
@@ -3996,14 +2403,6 @@ public class CamelMonitor extends CamelCommand {
             screenshotMessage = "Screenshot failed: " + e.getMessage();
             screenshotMessageTime = System.currentTimeMillis();
         }
-    }
-
-    private static String objToString(Object o) {
-        return o != null ? o.toString() : "";
-    }
-
-    private static long objToLong(Object o) {
-        return TuiHelper.objToLong(o);
     }
 
     record KeyRecord(String label, long timestamp) {
@@ -4180,15 +2579,6 @@ public class CamelMonitor extends CamelCommand {
         SelectionContext popup = actionsPopup.getSelectionContext();
         if (popup != null) {
             return popup;
-        }
-        if (tabsState.selected() == TAB_OVERVIEW) {
-            List<IntegrationInfo> infos = sortedOverviewInfos();
-            if (infos.isEmpty()) {
-                return null;
-            }
-            List<String> items = infos.stream().map(i -> i.name != null ? i.name : i.pid).toList();
-            Integer sel = overviewTableState.selected();
-            return new SelectionContext("table", items, sel != null ? sel : -1, items.size(), "Integrations");
         }
         MonitorTab tab = activeTab();
         return tab != null ? tab.getSelectionContext() : null;
